% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbIctpdDataArgs}
\alias{createGetDbIctpdDataArgs}
\title{Create a parameter object for the function getDbIctpdData}
\usage{
createGetDbIctpdDataArgs(drugTypeConceptIdList = c(38000182),
  conditionTypeConceptIdList = c(38000247), controlPeriodStart = -1080,
  controlPeriodEnd = -361, riskPeriodStart = 1, riskPeriodEnd = 30,
  censor = FALSE)
}
\arguments{
\item{drugTypeConceptIdList}{Which drug_type to use: generally only use 1 value (ex: 30dera).}

\item{controlPeriodStart}{start of the control period - can be set between -99999 and0,
default is -1080.}

\item{controlPeriodEnd}{end of the control period - can be set between -99999 and0,
default is -361.}

\item{riskPeriodStart}{start of the risk period - can be set between 0 and
99999,default is 1.}

\item{riskPeriodEnd}{end of the risk period - can be set between 0 and 99999,default
is 30.}

\item{censor}{a flag indicating wether the method should censor theobservation
period at the end of exposure or not. Availableinput is 0 or 1
with default = 0.}
}
\description{
Create a parameter object for the function getDbIctpdData
}
\details{
Create an object defining the parameter values.
}

