% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chronograph.R
\name{getChronographData}
\alias{getChronographData}
\title{Get the data for chronographs from the server.}
\usage{
getChronographData(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = NULL, cdmVersion = "5", exposureIds = c(),
  outcomeIds = c(), exposureOutcomePairs = NULL,
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era", shrinkage = 0.5,
  icPercentile = 0.025)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Name of database schema that contains OMOP CDM and vocabulary.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want all
temporary tables to be managed. Requires create/insert permissions
to this database.}

\item{cdmVersion}{Define the OMOP CDM version used: currently supports "5".}

\item{exposureIds}{A vector of IDs identifying the exposures to include when computing 
the expected count, i.e. the supplied IDs will define the 
background population in the chronograph. If the
exposure table is the drug_era table, these IDs correspond to
ingredient concept IDs. If the exposure table has the format of the
cohort table, these IDs correspond to the cohort definition IDs. If
left empty, all records in the exposure table will be used.}

\item{outcomeIds}{A vector of IDs identifying the outcomes to include when computing 
the expected count, i.e. the supplied IDs will define the 
background population in the chronograph. If the outcome
table is the drug_era table, these IDs correspond to condition
concept IDs. If the outcomes table has the format of the cohort
table, these IDs correspond to the cohort definition IDs. If left
empty, all records in the outcome table will be used.}

\item{exposureOutcomePairs}{A data frame with at least two columns:
         \itemize{
           \item {"exposureId" containing the drug_concept_ID or
                 cohort_concept_id of the exposure variable}
           \item {"outcomeId" containing the condition_concept_ID or
                 cohort_concept_id of the outcome variable}
         } Each row specifies an exposure-outcome-combination for 
which to collect data for. If left empty, all possible combinations of exposures and
outcomes will be computed, which could be time consuming.
To collect data for plotting multiple exposures and/or multiple outcomes in a single
 chronograph, two other "grouping" columns might be passed: 
         \itemize{
           \item {"exposureGrouping" with the same valueat each row with an exposure to
                 include in an exposure grouping}
           \item {"outcomeGrouping" with the same valueat each row with an exposure to
                 include in an exposure grouping}
}See details and examples below.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where the
exposure data is available.  If exposureTable = DRUG_ERA,
exposureSchema is not used by assumed to be cdmSchema.  Requires
read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If exposureTable
<> DRUG_ERA, then expectation is exposureTable has format of COHORT
table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where the data
used to define the outcome cohorts is available. If exposureTable =
CONDITION_ERA, exposureSchema is not used by assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If outcomeTable <>
CONDITION_OCCURRENCE, then expectation is outcomeTable has format of
COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{shrinkage}{Shrinkage used in IRR calculations, required >0 to deal with 0 case
counts, but larger number means more shrinkage. default is 0.5}

\item{icPercentile}{The lower bound of the credibility interval for the IC values
(IClow). default is 0.025,}
}
\value{
A data frame with observed and expected outcome counts in periods relative to the exposure
initiation date, for each outcome and exposure.
}
\description{
Get the data for creating chronographs from the server.
}
\details{
When mapping vocabularies e.g. medDRA to SNOMED, it is possible that multiple terms are found. To create a chronograph for the joint SNOMED-terms, 
          data collected from server for single exposures and outcomes can not be used for multiple exposures or multiple outcomes, 
          as the same person might be counted multiple times. You should instead specify grouping columns in the 
          exposureOutcomePairs-argument table.
}
\examples{
  # Create exposureOutcomePairs for multiple exposures and multiple outcomes, as well as for single rows

grouping_df <- cbind.data.frame("exposureGrouping"=1, "outcomeGrouping"=1, setNames(expand.grid(c(40170549, 1321636), c(321389,   73609,  134401)), c("exposureId", "outcomeId")))
single_df <- cbind.data.frame("exposureGrouping"=2:3, "outcomeGrouping"=2:3, "exposureId"=c(40170549, 1321636), "outcomeId"=c(321389,   73609))
exposureOutcomePairs <- rbind.data.frame(grouping_df, single_df)
          
 output <- getChronographData(connectionDetails = connectionDetails,
                              oracleTempSchema = NULL,
                              cdmDatabaseSchema = "synpuf5pct_20180710",
                              cdmVersion = "5",
                              exposureOutcomePairs = exposureOutcomePairs,
                              exposureDatabaseSchema = cdmDatabaseSchema,
                              exposureTable = "drug_era",
                              outcomeDatabaseSchema = cdmDatabaseSchema,
                              outcomeTable = "condition_era",
                              shrinkage = 0.5,
                              icPercentile = 0.025)#
 
 plot(IcTemporalPatternDiscovery::plotChronograph(data=output, exposureGrouping = 1, outcomeGrouping = 1, title="Exposures 40170549, 1321636 vs Outcomes 321389, 73609, 134401"))

}
