% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTPD.R
\docType{package}
\name{IcTemporalPatternDiscovery-package}
\alias{IcTemporalPatternDiscovery}
\alias{IcTemporalPatternDiscovery-package}
\title{IcTemporalPatternDiscovery: IC Temporal Pattern Discovery}
\description{
Population-level estimation method that estimates risk by combining a self-controlled and cohort design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/IcTemporalPatternDiscovery}
  \item \url{https://github.com/OHDSI/IcTemporalPatternDiscovery}
  \item Report bugs at \url{https://github.com/OHDSI/IcTemporalPatternDiscovery/issues}
}

}
\author{
\strong{Maintainer}: Tomas Bergvall \email{tomas.bergvall@who-umc.org}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
