% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICTPD.R
\name{ictpd}
\alias{ictpd}
\title{IC temporal pattern discovery}
\usage{
ictpd(connectionDetails, cdmSchema, resultsSchema, resultsTablePrefix = "ictpd", createResultsTable = TRUE, sourceName = "", exposureOutcomePairs, exposureTable = "drug_era", outcomeTable = "condition_era")
ictpd(connectionDetails, cdmSchema, resultsSchema, resultsTablePrefix = "ictpd", createResultsTable = TRUE, sourceName = "", exposureOutcomePairs, exposureTable = "drug_era", outcomeTable = "condition_era", analysisId = 1, drugTypeConceptIdList = c(38000182), conditionTypeConceptIdList = c(38000247))
ictpd(connectionDetails, cdmSchema, resultsSchema, resultsTablePrefix = "ictpd", createResultsTable = TRUE, sourceName = "", exposureOutcomePairs, exposureTable = "drug_era", outcomeTable = "condition_era", analysisId = 1, drugTypeConceptIdList = c(38000182), conditionTypeConceptIdList = c(38000247), controlPeriodStart = -1080, controlPeriodEnd = -361, multipleControlPeriods = 110, multipeObservationPeriods = '10000', shrinkage = 0.50, icPercentile = 0.025, metric = "IC025", censor = 0)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmSchema}{Name of database schema that contains OMOP CDM and vocabulary.}

\item{resultsSchema}{Name of database schema that we can write results to.}

\item{resultsTablePrefix}{Prefix used for the result tables in the \code{resultsSchema}.}

\item{createResultsTable}{If true, a new empty table will be created to store the results. If false, results will be inserted into the existing table.}

\item{sourceName}{Name of the database, as recorded in results.}

\item{exposureOutcomePairs}{A data frame with at least two columns:
\itemize{
  \item{"exposureConceptId" containing the drug_concept_ID or cohort_concept_id of the exposure variable}
  \item{"outcomeConceptId" containing the condition_concept_ID or cohort_concept_id of the outcome variable}
}}

\item{exposureTable}{\code{exposureTable="drugEra"} or \code{exposureTable="cohort"}.}

\item{outcomeTable}{\code{outcomeTable="conditionEra"} or \code{outcomeTable="cohort"}.}

\item{analysisId}{A unique identifier that can later be used to identify the results of this analysis.}

\item{drugTypeConceptIdList}{Which drug_type to use:  generally only use 1 value (ex:  30d era).}

\item{conditionTypeConceptIdList}{Which condition_type to use:  generally only use 1 value (ex:  30d era).}

\item{shrinkage}{shrinkage used in IRR calculations, required >0 to deal with 0 case counts, but larger number means more shrinkage. default is 0.5}

\item{ictpdAnalysisDetails}{object specifyng a set of analysis choices.}

\item{ictpdAnalysesDetails}{object specifyng one or several sets of analysis choices.}

\item{controlPeriodStart}{start of the control period - can be set between -99999 and 0, default is -1080}

\item{controlPeriodEnd}{end of the control period - can be set between -99999 and 0, default is -361}

\item{multipleControlPeriods}{Defines the control periods to use where 100 means the control period defined by controlPeriodStart/End, 010 means the period -30 to -1 day before prescription and 001 means the contorl period on the day of prescription}

\item{multipleObservationPeriods}{Defines the observation periods to use 10000 is 1-30 days, 01000 is 1 to 360 days, 00100 is 31 to 90 days, 00010 is 91 to 180 and 00001 is 721 to 1080 days after prescription default is '10000'}

\item{icPercentile}{the lower bound of the credibility interval for the IC values (IClow). default is 0.025,}

\item{metric}{defines wether the output will contain the point estimate or the lower bound. Available input is 'IC and 'IC025' default is 'IC025'}

\item{censor}{a flag indicating wether the method should censor the observation period at the end of exposure or not. Available input is 0 or 1 with default = 0.}
}
\value{
An object of type \code{ictpdResults} containing details for connecting to the database containing the results
}
\description{
\code{Ictpd} generates population-level estimation from OMOP CDMv4 instance by combining a self-controlled design with a cohort design.
}
\details{
Population-level estimation method that estimates risk by combining a self-controlled and cohort design.
}
\examples{
\dontrun{

connectionDetails <- createConnectionDetails(dbms="sql server", server="server_ip", schema="omop_cdm_database")
exposureOutcomePairs = data.frame(outcomeConceptId = c(196794, 196794, 312648), exposureConceptId = c(1501700, 1545958, 1551803))
ictpdResult <- ictpd(connectionDetails, cdmSchema = "OmopCdm_Simulated", resultsSchema = "OmopCdm_Simulated", sourceName = "OmopCdm_Simulated", exposureOutcomePairs = exposureOutcomePairs)

}
}
\references{
Norén GN, Bate A, Hopstadius J, Star K, Edwards IR.  Temporal Pattern Discovery for Trends and Transient Effect: Its Application to Patient Reocrds. by combining a self-controlled design with a cohort design
In: Proceedings of the fourteenth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (KDD'08).ACM Press, New York, pp 963-971
}

