% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chronograph.R
\name{plotChronograph}
\alias{plotChronograph}
\title{Plot a chronograph}
\usage{
plotChronograph(data, exposureId, outcomeId, exposureGrouping = NULL,
  outcomeGrouping = NULL, title = NULL, fileName = NULL)
}
\arguments{
\item{data}{Data as generated using the \code{getChronographData} function.}

\item{exposureId}{The unique ID identifying the exposure to plot.}

\item{outcomeId}{The unique ID identifying the outcome to plot.}

\item{title}{The title to show above the plot.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\description{
Creates a plot showing the observed and expected nunber of outcomes for each month in the 3 years
before and after initiation of the exposure, as well as the IC. The full process is described in
Noren et al.
}
\references{
Noren GN, Hopstadius J, Bate A, Star K, Edwards R, Temporal pattern discovery in longitudinal
electronic patient records, Data Mining and Knowledge Discovery, May 2010, Volume 20, Issue 3, pp
361-387.
}
