% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICTPD.R
\name{readIctpdAnalysesDetailsFromFile}
\alias{readIctpdAnalysesDetailsFromFile}
\title{readIctpdAnalysesDetailsFromFile}
\usage{
readIctpdAnalysesDetailsFromFile(file)
}
\arguments{
\item{file}{the name of the file to be loaded}
}
\value{
An object of type \code{analysesDetails}
}
\description{
\code{readIctpdAnalysesDetailsFromFile} reads an object of type \code{analysesDetails} from a CSV file
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="sql server", server="RNDUSRDHIT07.jnj.com")
  analysesDetails <- readIctpdAnalysesDetailsFromFile("c:/temp/test.csv")
  ictpdResult <- IcTemporalPatternDiscovery(analysesDetails, connectionDetails, "cdm_truven_mdcr", "scratch", sourceName = "cdm_truven_mdcr", exposuresOfInterest = c(767410,1314924,907879), outcomesOfInterest = c(444382, 79106, 138825), outcomeTable = "condition_era")
}
}

