% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICTPD.R
\name{runIctpd}
\alias{runIctpd}
\title{IC temporal pattern discovery}
\usage{
runIctpd(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  exposureOutcomePairs, exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era", outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era", drugTypeConceptIdList = c(),
  conditionTypeConceptIdList = c(), analysisId = 1,
  storeResultsInDatabase = FALSE, createOutputTables = TRUE,
  outputTablePrefix = "ictpd", controlPeriodStart = -1080,
  controlPeriodEnd = -361, riskPeriodStart = 1, riskPeriodEnd = 30,
  censor = FALSE, multipleControlPeriods = "110",
  multipleRiskPeriods = "10000", shrinkage = 0.5, icPercentile = 0.025,
  metric = "IC025")
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Name of database schema that contains OMOP CDM and vocabulary.}

\item{resultsDatabaseSchema}{Name of database schema that we can write results to.}

\item{exposureOutcomePairs}{A data frame with at least two columns:
\itemize{
  \item{"exposureConceptId" containing the drug_concept_ID or cohort_concept_id of the exposure variable}
  \item{"outcomeConceptId" containing the condition_concept_ID or cohort_concept_id of the outcome variable}
}}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where the exposure data is available.  If exposureTable = DRUG_ERA, exposureSchema is not used by assumed to be cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If exposureTable <> DRUG_ERA, then expectation is exposureTable has format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where the data used to define the outcome cohorts is available.  If exposureTable = CONDITION_ERA, exposureSchema is not used by assumed to be cdmSchema.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If outcomeTable <> CONDITION_OCCURRENCE, then expectation is outcomeTable has format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{drugTypeConceptIdList}{Which drug_type to use:  generally only use 1 value (ex:  30d era).}

\item{conditionTypeConceptIdList}{Which condition_type to use:  generally only use 1 value (ex:  30d era).}

\item{analysisId}{A unique identifier that can later be used to identify the results of this analysis.}

\item{storeResultsInDatabase}{Should the results be stored in the database?}

\item{createOutputTables}{Should the output tables be created? If not, they are assumed to exist and data will be appended.
If the value is true and the tables exist they will be overwritten.}

\item{outputTablePrefix}{Prefix used for the result tables in the \code{resultsDatabaseSchema}.}

\item{controlPeriodStart}{start of the control period - can be set between -99999 and 0, default is -1080.}

\item{controlPeriodEnd}{end of the control period - can be set between -99999 and 0, default is -361.}

\item{riskPeriodStart}{start of the risk period - can be set between 0 and 99999, default is 1.}

\item{riskPeriodEnd}{end of the risk period - can be set between 0 and 99999, default is 30.}

\item{censor}{a flag indicating wether the method should censor the observation period at the end of exposure or not. Available input is 0 or 1 with default = 0.}

\item{multipleControlPeriods}{Defines the control periods to use where 100 means the control period defined by controlPeriodStart/End, 010 means the period -30 to -1 day before prescription and 001 means the control period on the day of prescription}

\item{multipleRiskPeriods}{Defines the risk periods to use 10000 is 1-30 days, 01000 is 1 to 360 days, 00100 is 31 to 90 days, 00010 is 91 to 180 and 00001 is 721 to 1080 days after prescription default is '10000'}

\item{shrinkage}{Shrinkage used in IRR calculations, required >0 to deal with 0 case counts, but larger number means more shrinkage. default is 0.5}

\item{icPercentile}{The lower bound of the credibility interval for the IC values (IClow). default is 0.025,}

\item{metric}{Defines wether the output will contain the point estimate or the lower bound. Available input is 'IC and 'IC025' default is 'IC025'}
}
\value{
An object of type \code{ictpdResults} containing the results.
}
\description{
\code{Ictpd} generates population-level estimation from OMOP CDMv4 instance by combining a self-controlled design with a cohort design.
}
\details{
Population-level estimation method that estimates risk by combining a self-controlled and cohort design.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms="sql server", server="server_ip")
ictpdResults <- runIctpd(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema, exposureOutcomePairs)
ictpdResults
}
}
\references{
Noren GN, Bate A, Hopstadius J, Star K, Edwards IR.  Temporal Pattern Discovery for Trends and Transient Effect: Its Application to Patient Reocrds. by combining a self-controlled design with a cohort design
In: Proceedings of the fourteenth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (KDD'08).ACM Press, New York, pp 963-971
}

