% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICTPD.R
\name{runIctpdAnalyses}
\alias{runIctpdAnalyses}
\title{run ICTPD on a list of analysis}
\usage{
runIctpdAnalyses(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  exposureOutcomePairs, exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era", outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era", drugTypeConceptIdList = c(38000182),
  conditionTypeConceptIdList = c(38000247), storeResultsInDatabase = FALSE,
  createOutputTables = TRUE, outputTablePrefix = "ictpd", ictpdAnalysisList)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Name of database schema that contains OMOP CDM and vocabulary.}

\item{resultsDatabaseSchema}{Name of database schema that we can write results to.}

\item{exposureOutcomePairs}{A data frame with at least two columns:
\itemize{
  \item{"exposureConceptId" containing the drug_concept_ID or cohort_concept_id of the exposure variable}
  \item{"outcomeConceptId" containing the condition_concept_ID or cohort_concept_id of the outcome variable}
}}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where the exposure data is available.  If exposureTable = DRUG_ERA, exposureSchema is not used by assumed to be cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If exposureTable <> DRUG_ERA, then expectation is exposureTable has format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where the data used to define the outcome cohorts is available.  If exposureTable = CONDITION_ERA, exposureSchema is not used by assumed to be cdmSchema.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If outcomeTable <> CONDITION_OCCURRENCE, then expectation is outcomeTable has format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{drugTypeConceptIdList}{Which drug_type to use:  generally only use 1 value (ex:  30d era).}

\item{conditionTypeConceptIdList}{Which condition_type to use:  generally only use 1 value (ex:  30d era).}

\item{storeResultsInDatabase}{Should the results be stored in the database?}

\item{createOutputTables}{Should the output tables be created? If not, they are assumed to exist and data will be appended.
If the value is true and the tables exist they will be overwritten.}

\item{outputTablePrefix}{Prefix used for the result tables in the \code{resultsDatabaseSchema}.}

\item{ictpdAnalysisList}{A list of objects of type \code{ictpdAnalysis}.}
}
\value{
An object of type \code{ictpdResults} containing the results.
}
\description{
This function will run the different ICTPD variations specified in the analysis list on all the exposure-
outcome pairs.
}
\details{
Population-level estimation method that estimates risk by combining a self-controlled and cohort design.
This function will call both the \code{\link{getDbIctpdData}} and \code{\link{calculateStatisticsIc}} functions.
}
\examples{
\dontrun{
#Constructing a list of analyses:
analysisList <- NULL

analysis <- createIctpdAnalysis(analysisId = 1, censor = TRUE)
analysisList <- appendToIctpdAnalysisList(analysis)

analysis <- createIctpdAnalysis(analysisId = 2, censor = FALSE)
analysisList <- appendToIctpdAnalysisList(analysis, analysisList)

#You can store the list:
writeIctpdAnalysisListToFile(analysisList,"ictpdAnalysisList.csv")

#And load it back later:
analysisList <- loadIctpdAnalysisList("ictpdAnalysisList.csv")

#You can run all analyses across a set of exposure-outcome pairs with one command:
connectionDetails <- createConnectionDetails(dbms="sql server", server="server_ip")
exposureOutcomePairs = data.frame(outcomeConceptId = c(196794, 196794, 312648),
                                  exposureConceptId = c(1501700, 1545958, 1551803))

results <- runIctpdAnalyses(connectionDetails,
                            cdmDatabaseSchema = "cdm_schema.dbo",
                            resultsDatabaseSchema = "results.dbo",
                            exposureOutcomePairs = exposureOutcomePairs)),
                            ictpdAnalysisList = ictpdAnalysisList)
}
}
\references{
Noren GN, Bate A, Hopstadius J, Star K, Edwards IR.  Temporal Pattern Discovery for Trends and Transient Effect: Its Application to Patient Reocrds. by combining a self-controlled design with a cohort design
In: Proceedings of the fourteenth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (KDD'08).ACM Press, New York, pp 963-971
}

