% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveIctpdAnalysisList}
\alias{saveIctpdAnalysisList}
\title{Save a list of ictpdAnalysis to file}
\usage{
saveIctpdAnalysisList(ictpdAnalysisList, file)
}
\arguments{
\item{ictpdAnalysisList}{The ictpdAnalysis list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{ictpdAnalysis} to file. The file is in JSON format.
}
