% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICTPD.R
\name{writeIctpdAnalysesDetailsToFile}
\alias{writeIctpdAnalysesDetailsToFile}
\title{writeIctpdAnalysesDetailsToFile}
\usage{
writeIctpdAnalysesDetailsToFile(ictpdAnalysesDetails, file)
}
\arguments{
\item{ictpdAnalysesDetails}{the \code{ictpdAnalysesDetails} to be written to file}

\item{file}{the name of the file where the results will be written}
}
\description{
\code{writeIctpdAnalysesDetailsToFile} writes an object of type \code{analysesDetails} to a CSV file
}
\examples{
\dontrun{
  analysesDetails <- NULL
  analysesDetails <- appendToIctpdAnalysesDetails(createIctpdAnalysisDetails(analysisId = 1),analysesDetails)
  analysesDetails <- appendToIctpdAnalysesDetails(createIctpdAnalysisDetails(analysisId = 2),analysesDetails)
  writeIctpdAnalysesDetailsToFile(analysesDetails,"c:/temp/test.csv")
}
}

