% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICTPDAnalysis.R
\name{writeIctpdAnalysisList}
\alias{writeIctpdAnalysisList}
\title{writeIctpdAnalysisListToFile}
\usage{
writeIctpdAnalysisList(ictpdAnalysisList, file)
}
\arguments{
\item{ictpdAnalysisList}{the \code{ictpdAnalysisList} to be written to file}

\item{file}{the name of the file where the results will be written}
}
\description{
\code{writeIctpdAnalysisListToFile} writes an object of type \code{IctpdAnalysisList} to a CSV file
}
\examples{
\dontrun{
#Constructing a list of analyses:
analysisList <- NULL

analysis <- createIctpdAnalysis(analysisId = 1, censor = TRUE)
analysisList <- appendToIctpdAnalysisList(analysis)

analysis <- createIctpdAnalysis(analysisId = 2, censor = FALSE)
analysisList <- appendToIctpdAnalysisList(analysis, analysisList)

#You can store the list:
writeIctpdAnalysisListToFile(analysisList,"ictpdAnalysisList.csv")

#And load it back later:
analysisList <- loadIctpdAnalysisList("ictpdAnalysisList.csv")

#You can run all analyses across a set of exposure-outcome pairs with one command:
connectionDetails <- createConnectionDetails(dbms="sql server", server="server_ip")
exposureOutcomePairs = data.frame(outcomeConceptId = c(196794, 196794, 312648),
                                  exposureConceptId = c(1501700, 1545958, 1551803))

results <- runIctpdAnalyses(connectionDetails,
                            cdmDatabaseSchema = "cdm_schema.dbo",
                            resultsDatabaseSchema = "results.dbo",
                            exposureOutcomePairs = exposureOutcomePairs)),
                            ictpdAnalysisList = ictpdAnalysisList)
}
}

