% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICTPDAnalysis.R
\name{createIctpdAnalysis}
\alias{createIctpdAnalysis}
\title{createSccAnalysisDetails}
\usage{
createIctpdAnalysis(analysisId = 1, controlPeriodStart = -1080,
  controlPeriodEnd = -361, riskPeriodStart = 1, riskPeriodEnd = 30,
  censor = FALSE, multipleControlPeriods = "110",
  multipleRiskPeriods = "10000", shrinkage = 0.5, icPercentile = 0.025,
  metric = "IC025")
}
\arguments{
\item{analysisId}{A unique identifier that can later be used to identify the results of this analysis}

\item{controlPeriodStart}{start of the control period - can be set between -99999 and 0, default is -1080.}

\item{controlPeriodEnd}{end of the control period - can be set between -99999 and 0, default is -361.}

\item{riskPeriodStart}{start of the risk period - can be set between 0 and 99999, default is 1.}

\item{riskPeriodEnd}{end of the risk period - can be set between 0 and 99999, default is 30.}

\item{censor}{a flag indicating wether the method should censor the observation period at the end of exposure or not. Available input is 0 or 1 with default = 0.}

\item{multipleControlPeriods}{Defines the control periods to use where 100 means the control period defined by controlPeriodStart/End, 010 means the period -30 to -1 day before prescription and 001 means the control period on the day of prescription}

\item{multipleRiskPeriods}{Defines the risk periods to use 10000 is 1-30 days, 01000 is 1 to 360 days, 00100 is 31 to 90 days, 00010 is 91 to 180 and 00001 is 721 to 1080 days after prescription default is '10000'}

\item{shrinkage}{Shrinkage used in IRR calculations, required >0 to deal with 0 case counts, but larger number means more shrinkage. default is 0.5}

\item{icPercentile}{The lower bound of the credibility interval for the IC values (IClow). default is 0.025,}

\item{metric}{Defines wether the output will contain the point estimate or the lower bound. Available input is 'IC and 'IC025' default is 'IC025'}
}
\description{
\code{createIctpdAnalysis} generates an object specifying one set of analysis choices
for the IC Temporal Pattern Discovery method.
}
\examples{
\dontrun{
#Constructing a list of analyses:
analysisList <- NULL

analysis <- createIctpdAnalysis(analysisId = 1, censor = TRUE)
analysisList <- appendToIctpdAnalysisList(analysis)

analysis <- createIctpdAnalysis(analysisId = 2, censor = FALSE)
analysisList <- appendToIctpdAnalysisList(analysis, analysisList)

#You can store the list:
writeIctpdAnalysisListToFile(analysisList,"ictpdAnalysisList.csv")

#And load it back later:
analysisList <- loadIctpdAnalysisList("ictpdAnalysisList.csv")

#You can run all analyses across a set of exposure-outcome pairs with one command:
connectionDetails <- createConnectionDetails(dbms="sql server", server="server_ip")
exposureOutcomePairs = data.frame(outcomeConceptId = c(196794, 196794, 312648),
                                  exposureConceptId = c(1501700, 1545958, 1551803))

results <- runIctpdAnalyses(connectionDetails,
                            cdmDatabaseSchema = "cdm_schema.dbo",
                            resultsDatabaseSchema = "results.dbo",
                            exposureOutcomePairs = exposureOutcomePairs)),
                            ictpdAnalysisList = ictpdAnalysisList)
}
}

