% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadIctpdAnalysisList}
\alias{loadIctpdAnalysisList}
\title{Load a list of ictpdAnalysis from file}
\usage{
loadIctpdAnalysisList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{ictpdAnalysis}.
}
\description{
Load a list of objects of type \code{ictpdAnalysis} from file. The file is in JSON format.
}

